refactor: move memory consolidation runner
