+++ src/memory/consolidation-runner.ts
